Add and refactor vm category
