+++ maro/simulator/scenarios/vm_scheduling/__init__.py
@@ -2,18 +2,18 @@
 # Licensed under the MIT license.
 
 from .business_engine import VmSchedulingBusinessEngine
-from .common import AllocateAction, DecisionPayload, Latency, PmState, PostponeAction, PostponeType, VmRequestPayload
+from .common import AllocateAction, DecisionPayload, Latency, PostponeAction, VmRequestPayload
 from .cpu_reader import CpuReader
-from .events import Events
+from .enums import Events, PmState, PostponeType, VmCategory
 from .physical_machine import PhysicalMachine
 from .virtual_machine import VirtualMachine
 
 __all__ = [
     "VmSchedulingBusinessEngine",
-    "AllocateAction", "PostponeAction", "DecisionPayload", "Latency", "PmState", "PostponeType", "VmCategory",
+    "AllocateAction", "PostponeAction", "DecisionPayload", "Latency",
     "VmRequestPayload",
     "CpuReader",
-    "Events",
+    "Events", "PmState", "PostponeType", "VmCategory",
     "PhysicalMachine",
     "VirtualMachine"
 ]
