--- maro/simulator/scenarios/vm_scheduling/events.py
@@ -1,10 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT license.
-
-from enum import Enum
-
-
-class Events(Enum):
-    """VM-PM pairs related events."""
-    # VM request events.
-    REQUEST = "vm_required"
